requirements.txt
contourpy==1.0.7
cycler==0.11.0
emoji==2.2.0
fonttools==4.39.2
importlib-resources==5.12.0
kaleido==0.2.1
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
regex==2022.10.31
six==1.16.0
tenacity==8.2.2
wordcloud==1.8.2.2
zipp==3.15.0
Jinja2==3.1.2
folium==0.14.0
python-whois==0.7.3
flask==2.0.2
flask_restx==0.5.1
werkzeug==2.1.2
fpdf2==2.7.3